requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.4.2
Django==1.11.12
django-cors-headers==2.4.0
djangorestframework==3.9.0
idna==2.7
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.7
six==1.11.0


# 批量导入虚拟环境所依赖的库
# pip install -r requirements.txt
# 批量导出虚拟环境所依赖的库
# pip freeze >requirements.txt